requirements.txt
Django==4.2.14
psycopg2==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8
django-model-utils==4.5.1
pyTelegramBotAPI==4.18.1
pillow==10.4.0
siwe==4.2.0
gunicorn==22.0.0
redis==5.0.4
celery==5.4.0
django-celery-beat==2.6.0

